A pane opens upon request
This commit creates two different 'classes' that deal with maintaining
tmux related information.  The first of these two classes is the 'main'
class, which is a general controller class and storage for information
about the panes that are created.

The second class is 'pane', which contains information about a pane that
has been opened.  The 'main' class should represent the current tmux
window, while having 'many' panes.

Actually being able to handle this madness while simultaneously allowing
the user to do things in panes that might end up closing them... might
not be a nightmare but might as well be.