Open panes with a command
The pane goes away as soon as the command is done.  This really doesn't
help us that much, and there is no real event api for tmux, I'm going to
have to build some tooling that will do some cleanup for tmux so that it
can remove windows from its system that no longer exist.

I have several ideas for how to do that, and none of them are
particularly awesome.