Delete externally destroyed panes.
Sometimes people are going to destroy panes outside of this system,
which were created by this system.  I see two potential cases for this
currently:

1. Someone runs a create pand with a very short command, like 'ls'.  The
pane will get destroyed before I can collect enough information about
the pane to know whether or not it is still there.

2. Someone will go into another pane and kill it, or use tmux commands
to blow it away manually.

In case #1, capture the issue and don't store the pane data.

In case #2, offer a command that allows clean up of panes lost this way.
In addition, call this method at the beginning of create pane.