Handle high numbered main panes
I wasn't doing string coercions to strings properly, so when I started
with pane %4, and got pane %22, the system would not correctly guess the
correct newest pane (it would assume it was 4, because it was the
'highest' numbered by string ordering).

This adjusts that so that it actually uses integer comparisons to find
the highest ordered pane.