Window concept
I could leave well enough alone and just control panes in the current
window, but that would really be selling the system short.  I want
people to be able to do far more if they want.

This introduces the concept of 'windows', which will eventually contain
the panes that I'm already controlling.  Code currently compiles, but is
not fully tested.