Correct active pane searches
The problem with almost all searches and updates in the tmux system is
that you have no idea what the window/pane relationship is going to be
by index at any given time.  Everything has 'id's' that can be searched
for, but the indexes can change, and everything is checked by index, not
by id.

Regardless, this updates active pane searches to use the window index
which is returned by pane listings.